+++ lib/hcrypto/bn.c
@@ -237,7 +237,7 @@ BN_is_bit_set(const BIGNUM *bn, int bit)
     heim_integer *hi = (heim_integer *)bn;
     unsigned char *p = hi->data;
 
-    if ((bit / 8) > hi->length || hi->length == 0)
+    if ((size_t)(bit / 8) > hi->length || hi->length == 0)
 	return 0;
 
     return p[hi->length - 1 - (bit / 8)] & is_set[bit % 8];
@@ -249,7 +249,7 @@ BN_set_bit(BIGNUM *bn, int bit)
     heim_integer *hi = (heim_integer *)bn;
     unsigned char *p;
 
-    if ((bit / 8) > hi->length || hi->length == 0) {
+    if ((size_t)(bit / 8) > hi->length || hi->length == 0) {
 	size_t len = (bit + 7) / 8;
 	void *d = realloc(hi->data, len);
 	if (d == NULL)
@@ -271,7 +271,7 @@ BN_clear_bit(BIGNUM *bn, int bit)
     heim_integer *hi = (heim_integer *)bn;
     unsigned char *p = hi->data;
 
-    if ((bit / 8) > hi->length || hi->length == 0)
+    if ((size_t)(bit / 8) > hi->length || hi->length == 0)
 	return 0;
 
     p[hi->length - 1 - (bit / 8)] &= (unsigned char)(~(is_set[bit % 8]));
@@ -309,7 +309,7 @@ BN_get_word(const BIGNUM *bn)
     if (hi->negative || hi->length > sizeof(num))
 	return ULONG_MAX;
 
-    for (i = 0; i < hi->length; i++)
+    for (i = 0; (size_t)i < hi->length; i++)
 	num = ((unsigned char *)hi->data)[i] | (num << 8);
     return num;
 }
@@ -336,7 +336,7 @@ BN_rand(BIGNUM *bn, int bits, int top, int bottom)
 
     {
 	size_t j = len * 8;
-	while(j > bits) {
+	while(j > (size_t)bits) {
 	    BN_clear_bit(bn, j - 1);
 	    j--;
 	}
