+++ lib/hcrypto/engine.c
@@ -377,7 +377,7 @@ ENGINE_by_id(const char *id)
 {
     int i;
 
-    for (i = 0; i < num_engines; i++) {
+    for (i = 0; (unsigned int)i < num_engines; i++) {
 	if (strcmp(id, engines[i]->id) == 0) {
 	    ENGINE_up_ref(engines[i]);
 	    return engines[i];
